# openapi.yaml
openapi: 3.0.3
info:
  title: Movie Reservation System
  version: 1.0.0
  description: |
    Every successful response has the shape `{ message, data }`.
    Every error response has the shape `{ error, message? }` — `error` is a stable error code, `message` only appears when extra explanation is needed.
servers:
  - url: http://localhost:4000
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
  schemas:
    Genre:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Sci-Fi }
        deletedAt: { type: string, format: date-time, nullable: true }
    Movie:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string, example: Blade Runner 2049 }
        description: { type: string }
        posterImageId: { type: string, nullable: true }
        posterUrl:
          type: string
          nullable: true
          description: Full image URL derived from posterImageId
          example: http://localhost:4000/api/image/poster-1234567890-123456789.jpg
        genreId: { type: string, format: uuid }
        genre: { $ref: '#/components/schemas/Genre' }
        deletedAt: { type: string, format: date-time, nullable: true }
...
    Showtime:
      type: object
      properties:
        id: { type: string, format: uuid }
        movieId: { type: string, format: uuid }
        theaterId: { type: string, format: uuid }
        startsAt: { type: string, format: date-time }
        price:
          type: string
          description: Decimal(10,2), serialized as a string
          example: '95000'
        movie:
          type: object
          properties:
            id: { type: string, format: uuid }
            title: { type: string }
        theater:
          type: object
          properties:
            id: { type: string, format: uuid }
            name: { type: string }
...
    Pagination:
      type: object
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 20 }
        total: { type: integer, example: 42 }
        totalPages: { type: integer, example: 3 }
    Error:
      type: object
      properties:
        error: { type: string, description: Stable error code, example: GenreNotFound }
        message: { type: string, description: Only present when extra explanation is needed }
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            issues:
              type: array
              description: Zod issue list, only present when error is ValidationError
              items: { type: object }
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationError' }
    Unauthorized:
      description: Missing cookie, invalid token, or a token revoked by logout
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: Unauthorized }
    Forbidden:
      description: Authenticated but not an ADMIN
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: Forbidden }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
security:
  - cookieAuth: []
paths:
...
  /api/movies:
    get:
      tags: [Movies]
      summary: List movies
      security: []
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: genreId, in: query, schema: { type: string, format: uuid } }
        - { name: search, in: query, description: 'Case-insensitive match on title', schema: { type: string }, example: godfather }
        - { name: orderBy, in: query, description: Ordered by creation time, schema: { type: string, enum: [asc, desc], default: desc } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data:
                    allOf:
                      - $ref: '#/components/schemas/Pagination'
                      - type: object
                        properties:
                          items:
                            type: array
                            items: { $ref: '#/components/schemas/Movie' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/movies/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Movies]
      summary: Get a movie
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Movie' }
        '404':
          description: MovieNotFound
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: MovieNotFound }
  /api/genres:
    get:
      tags: [Genres]
      summary: List genres
      security: []
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: orderBy, in: query, description: Ordered by name, schema: { type: string, enum: [asc, desc], default: desc } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data:
                    allOf:
                      - $ref: '#/components/schemas/Pagination'
                      - type: object
                        properties:
                          items:
                            type: array
                            items: { $ref: '#/components/schemas/Genre' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/genres/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Genres]
      summary: Get a genre
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Genre' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/showtimes:
    get:
      tags: [Showtimes]
      summary: List showtimes
      security: []
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: movieId, in: query, schema: { type: string, format: uuid } }
        - { name: theaterId, in: query, schema: { type: string, format: uuid } }
        - { name: from, in: query, description: 'startsAt >= from; accepts YYYY-MM-DD or an ISO date-time', schema: { type: string }, example: '2026-10-01' }
        - { name: to, in: query, description: 'startsAt <= to; accepts YYYY-MM-DD or an ISO date-time', schema: { type: string }, example: '2026-10-31' }
        - { name: orderBy, in: query, description: Ordered by startsAt, schema: { type: string, enum: [asc, desc], default: asc } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data:
                    allOf:
                      - $ref: '#/components/schemas/Pagination'
                      - type: object
                        properties:
                          items:
                            type: array
                            items: { $ref: '#/components/schemas/Showtime' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/showtimes/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Showtimes]
      summary: Get a showtime
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Showtime' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/admin/movies:
    post:
      tags: [Admin]
      summary: Create a movie
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [title, description, genreId, poster]
              properties:
                title: { type: string, example: Blade Runner 2049 }
                description: { type: string, example: A young blade runner uncovers a secret that could tip the balance of society. }
                genreId: { type: string, format: uuid }
                poster: { type: string, format: binary, description: 'jpeg/png/gif, up to 10MB' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Movie' }
        '400':
          description: '`PosterRequired`, `GenreNotFound` (genre missing or soft-deleted), `UploadError`, or `ValidationError`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: PosterRequired, message: Poster image is required }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/admin/movies/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [Admin]
      summary: Update a movie — at least one field or a poster
      description: When the poster changes, the previous file is deleted from disk after a successful update.
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title: { type: string, example: "Blade Runner 2049 (Director's Cut)" }
                description: { type: string }
                genreId: { type: string, format: uuid }
                poster: { type: string, format: binary }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Movie' }
        '400':
          description: '`EmptyUpdate` or `GenreNotFound`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: EmptyUpdate, message: At least one field or a poster is required }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Admin]
      summary: Soft-delete a movie
      responses:
        '200': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/admin/genres:
    post:
      tags: [Admin]
      summary: Create a genre
      description: If the name matches a soft-deleted genre, that genre is restored instead of returning an error.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, example: Documentary }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Genre' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: GenreAlreadyExists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: GenreAlreadyExists }
  /api/admin/genres/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [Admin]
      summary: Rename a genre
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, example: Docu-drama }
      responses:
        '200': { description: Updated }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: GenreAlreadyExists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [Admin]
      summary: Soft-delete a genre
      description: Refused while any non-deleted movie still uses this genre.
      responses:
        '200': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: GenreInUse
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: GenreInUse, message: 3 movie(s) still use this genre }
...
  /api/admin/showtimes:
    post:
      tags: [Admin]
      summary: Create a showtime
      description: A theater cannot have two showtimes with the same start time.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [movieId, theaterId, startsAt, price]
              properties:
                movieId: { type: string, format: uuid }
                theaterId: { type: string, format: uuid }
                startsAt: { type: string, format: date-time, example: '2026-10-01T19:00:00Z' }
                price: { type: number, minimum: 0, exclusiveMinimum: true, example: 95000 }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Showtime' }
        '400':
          description: '`MovieNotFound` or `TheaterNotFound` (missing or soft-deleted)'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: MovieNotFound }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: ShowtimeConflict — the theater already has a showtime at this exact start time
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: ShowtimeConflict }
  /api/admin/showtimes/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [Admin]
      summary: Update a showtime — at least one field
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                movieId: { type: string, format: uuid }
                theaterId: { type: string, format: uuid }
                startsAt: { type: string, format: date-time }
                price: { type: number, example: 120000 }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Showtime' }
        '400':
          description: '`EmptyUpdate`, `MovieNotFound` or `TheaterNotFound`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: EmptyUpdate, message: At least one field is required }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: ShowtimeConflict
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [Admin]
      summary: Delete a showtime
      description: A hard delete, not a soft delete. Refused once the showtime has reservations.
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Showtime' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: ShowtimeHasReservations
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: ShowtimeHasReservations, message: 2 reservation(s) already booked }
...